NO-JIRA: Explictily convert double to uint64_t in qpid-txtest2 to avoid warnings on some compilers.
git-svn-id: https://svn.apache.org/repos/asf/qpid/trunk/qpid@1622372 13f79535-47bb-0310-9956-ffa450edef68

+++ cpp/src/tests/qpid-txtest2.cpp
@@ -201,7 +201,7 @@ struct Transfer : public TransactionalClient, public Runnable
                 id << source << ">" << target << ":" << t+1;
                 try {
                     for (uint m = 0; m < opts.msgsPerTx; m++) {
-                        Message msg = receiver.fetch(Duration::SECOND*opts.fetchTimeout);
+                        Message msg = receiver.fetch(Duration::SECOND*uint64_t(opts.fetchTimeout));
                         if (msg.getContentSize() != opts.size) {
                             std::ostringstream oss;
                             oss << "Message size incorrect: size=" << msg.getContentSize() << "; expected " << opts.size;
